requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
pillow>=10.0.0
tqdm>=4.66.0
scipy>=1.11.0
transformers>=4.40.0
joblib>=1.3.0
opencv-python-headless>=4.9.0
